% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfbseedR-package.R
\docType{package}
\name{cfbseedR-package}
\alias{cfbseedR}
\alias{cfbseedR-package}
\title{cfbseedR: Functions to Efficiently Simulate and Evaluate College Football Seasons}
\description{
A set of functions to simulate college football seasons and compute standings, conference ranks, conference champions, and College Football Playoff seeds. The package is an adaptation of the 'nflseedR' package by Sebastian Carl and Lee Sharpe (MIT licensed) with college football semantics: conferences instead of divisions, independents, conference championship games, and 12-team playoff straight seeding.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sportsdataverse/cfbseedR}
  \item \url{https://cfbseedR.sportsdataverse.org}
  \item Report bugs at \url{https://github.com/sportsdataverse/cfbseedR/issues}
}

}
\author{
\strong{Maintainer}: Saiem Gilani \email{saiem.gilani@gmail.com} (Adapted from nflseedR by Sebastian Carl and Lee Sharpe)

}
\keyword{internal}
